Create RandomBarcode.py
This program generates a random 12-digit number using the (random) library. It then uses the (code128) library to generate a barcode image from this number. The barcode image is saved as a PNG file using the (save) method. Finally, the program prints a message indicating that the barcode has been saved.